# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "mlndex-backend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/translations/{id}": {
      "get": {
        "tags": [
          "Translations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Translations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}",
            "text/json": "@{schema=}",
            "application/*+json": "@{schema=}"
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Translations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/translation-teams": {
      "post": {
        "tags": [
          "TranslationTeams"
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}",
            "text/json": "@{schema=}",
            "application/*+json": "@{schema=}"
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "TranslationTeams"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": "@{text/plain=; application/json=; text/json=}"
          }
        }
      }
    },
    "/api/translation-teams/{id}/members": {
      "get": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": "@{text/plain=; application/json=; text/json=}"
          }
        }
      }
    },
    "/api/translation-teams/my-teams": {
      "get": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": "@{type=integer; format=int32; default=5}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/translation-teams/{id}": {
      "put": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}",
            "text/json": "@{schema=}",
            "application/*+json": "@{schema=}"
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/translation-teams/{id}/invitations": {
      "post": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}",
            "text/json": "@{schema=}",
            "application/*+json": "@{schema=}"
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/translation-teams/invitations/{invitationId}/accept": {
      "post": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/translation-teams/invitations/{invitationId}/reject": {
      "post": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/translation-teams/{id}/join-requests": {
      "post": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}",
            "text/json": "@{schema=}",
            "application/*+json": "@{schema=}"
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/translation-teams/join-requests/{requestId}/approve": {
      "post": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/translation-teams/join-requests/{requestId}/reject": {
      "post": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/translation-teams/{id}/members/{userId}": {
      "delete": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/translation-teams/{id}/members/{userId}/role": {
      "put": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": "@{schema=}",
            "text/json": "@{schema=}",
            "application/*+json": "@{schema=}"
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/translation-teams/{id}/series": {
      "get": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/translation-teams/{id}/stats": {
      "get": {
        "tags": [
          "TranslationTeams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": "@{type=integer; format=int32}"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}